Many minor improvements to providers. Created static provider for Old RMC and PMC like usage.

+++ Source/RuntimeMeshComponent/Private/Providers/RuntimeMeshProviderBox.cpp
@@ -36,7 +36,7 @@ void FRuntimeMeshProviderBoxProxy::Initialize()
 	CreateSection(0, 0, Properties);
 }
 
-bool FRuntimeMeshProviderBoxProxy::GetSectionMeshForLOD(uint8 LODIndex, int32 SectionId, FRuntimeMeshRenderableMeshData& MeshData)
+bool FRuntimeMeshProviderBoxProxy::GetSectionMeshForLOD(int32 LODIndex, int32 SectionId, FRuntimeMeshRenderableMeshData& MeshData)
 {
 	// We should only ever be queried for section 0 and lod 0
 	check(SectionId == 0 && LODIndex == 0);
@@ -136,9 +136,7 @@ FRuntimeMeshCollisionSettings FRuntimeMeshProviderBoxProxy::GetCollisionSettings
 	Settings.bUseComplexAsSimple = false;
 
 	Settings.Boxes.Emplace(BoxRadius.X * 2, BoxRadius.Y * 2, BoxRadius.Z * 2);
-
-
-
+	   	 
 	return Settings;
 }
